refactor(params): reorganise param components dir structure
