+++ app/components/lookbook/params/editor/component.js
