+++ app/components/lookbook/params/editor/component.rb
@@ -1,0 +1,13 @@
+module Lookbook
+  module Params
+    class Editor::Component < Lookbook::BaseComponent
+      renders_many :fields, Lookbook::Params::Field::Component
+
+      protected
+
+      def alpine_component
+        "paramsEditorComponent"
+      end
+    end
+  end
+end
